--- a/.github/CODEOWNERS
+++ b/.github/CODEOWNERS
@@ -93,6 +93,7 @@
 /src/stash-debug                    @MaterializeInc/adapter
 /src/storage                        @MaterializeInc/storage
 /src/storage-client                 @MaterializeInc/storage
+/src/storage-operators              @MaterializeInc/storage
 /src/storage-types                  @MaterializeInc/storage
 /src/test-macro                     @MaterializeInc/testing
 /src/testdrive                      @MaterializeInc/testing