--- a/src/adapter/src/catalog/builtin_table_updates.rs
+++ b/src/adapter/src/catalog/builtin_table_updates.rs
@@ -32,10 +32,10 @@ use mz_sql::catalog::{CatalogCluster, CatalogDatabase, CatalogSchema, CatalogTyp
 use mz_sql::func::FuncImplCatalogDetails;
 use mz_sql::names::{CommentObjectId, ResolvedDatabaseSpecifier, SchemaId, SchemaSpecifier};
 use mz_sql_parser::ast::display::AstDisplay;
-use mz_storage_client::types::connections::inline::ReferencedConnection;
-use mz_storage_client::types::connections::KafkaConnection;
-use mz_storage_client::types::sinks::{KafkaSinkConnection, StorageSinkConnection};
-use mz_storage_client::types::sources::{
+use mz_storage_types::connections::inline::ReferencedConnection;
+use mz_storage_types::connections::KafkaConnection;
+use mz_storage_types::sinks::{KafkaSinkConnection, StorageSinkConnection};
+use mz_storage_types::sources::{
     GenericSourceConnection, KafkaSourceConnection, PostgresSourceConnection,
 };
 
@@ -554,18 +554,16 @@ impl CatalogState {
                 Datum::String(&schema_id.to_string()),
                 Datum::String(name),
                 Datum::String(match connection.connection {
-                    mz_storage_client::types::connections::Connection::Kafka { .. } => "kafka",
-                    mz_storage_client::types::connections::Connection::Csr { .. } => {
+                    mz_storage_types::connections::Connection::Kafka { .. } => "kafka",
+                    mz_storage_types::connections::Connection::Csr { .. } => {
                         "confluent-schema-registry"
                     }
-                    mz_storage_client::types::connections::Connection::Postgres { .. } => {
-                        "postgres"
-                    }
-                    mz_storage_client::types::connections::Connection::Aws(..) => "aws",
-                    mz_storage_client::types::connections::Connection::AwsPrivatelink(..) => {
+                    mz_storage_types::connections::Connection::Postgres { .. } => "postgres",
+                    mz_storage_types::connections::Connection::Aws(..) => "aws",
+                    mz_storage_types::connections::Connection::AwsPrivatelink(..) => {
                         "aws-privatelink"
                     }
-                    mz_storage_client::types::connections::Connection::Ssh { .. } => "ssh-tunnel",
+                    mz_storage_types::connections::Connection::Ssh { .. } => "ssh-tunnel",
                 }),
                 Datum::String(&owner_id.to_string()),
                 privileges,
@@ -573,7 +571,7 @@ impl CatalogState {
             diff,
         }];
         match connection.connection {
-            mz_storage_client::types::connections::Connection::Ssh(ref ssh) => {
+            mz_storage_types::connections::Connection::Ssh(ref ssh) => {
                 if let Some(public_key_set) = ssh.public_keys.as_ref() {
                     updates.extend(self.pack_ssh_tunnel_connection_update(
                         id,
@@ -584,13 +582,13 @@ impl CatalogState {
                     tracing::error!("does this even happen?");
                 }
             }
-            mz_storage_client::types::connections::Connection::Kafka(ref kafka) => {
+            mz_storage_types::connections::Connection::Kafka(ref kafka) => {
                 updates.extend(self.pack_kafka_connection_update(id, kafka, diff));
             }
-            mz_storage_client::types::connections::Connection::Csr(_)
-            | mz_storage_client::types::connections::Connection::Postgres(_)
-            | mz_storage_client::types::connections::Connection::Aws(_)
-            | mz_storage_client::types::connections::Connection::AwsPrivatelink(_) => {
+            mz_storage_types::connections::Connection::Csr(_)
+            | mz_storage_types::connections::Connection::Postgres(_)
+            | mz_storage_types::connections::Connection::Aws(_)
+            | mz_storage_types::connections::Connection::AwsPrivatelink(_) => {
                 if let Some(aws_principal_context) = self.aws_principal_context.as_ref() {
                     updates.extend(self.pack_aws_privatelink_connection_update(
                         id,