--- a/src/adapter/src/coord/ddl.rs
+++ b/src/adapter/src/coord/ddl.rs
@@ -94,7 +94,7 @@ impl Coordinator {
     /// function successfully returns on any built `DataflowDesc`.
     ///
     /// [`CatalogState`]: crate::catalog::CatalogState
-    /// [`DataflowDesc`]: mz_compute_client::types::dataflows::DataflowDesc
+    /// [`DataflowDesc`]: mz_compute_types::dataflows::DataflowDesc
     #[tracing::instrument(level = "debug", skip_all)]
     pub(crate) async fn catalog_transact_with<'a, F, R>(
         &mut self,