… Cargo.lock
… Cargo.toml
… ci/test/cargo-test-miri-fast.sh
… ci/test/cargo-test-miri.sh
… doc/developer/design/20230802_explain_running_dataflows.md
… src/adapter/Cargo.toml
… src/adapter/src/coord/timeline.rs
… src/adapter/src/coord/timestamp_selection.rs
… src/adapter/src/error.rs
… src/adapter/src/explain/lir/mod.rs
… src/adapter/src/explain/mir/mod.rs
… src/adapter/src/explain/mod.rs
… src/adapter/src/explain/optimizer_trace.rs
… src/adapter/src/flags.rs
… src/adapter/src/session.rs
… src/adapter/src/subscribe.rs
… src/adapter/src/util.rs
… src/adapter/tests/timestamp_selection.rs
… src/cluster-client/src/client.rs
… src/cluster/Cargo.toml
… src/cluster/src/server.rs
… src/cluster/src/types.rs
… src/clusterd/Cargo.toml
… src/clusterd/src/bin/clusterd.rs
… src/compute-client/Cargo.toml
… src/compute-client/build.rs
… src/compute-client/src/controller.rs
… src/compute-client/src/controller/instance.rs
… src/compute-client/src/logging.proto
… src/compute-client/src/metrics.rs