This allows changes to implementation code to invalidate fewer crates, hopefully improving dev iteration times.

Notably, mz-sql and mz-transform (both heavyweight) no longer depend on storage-client. As mentioned in the reviewer notes, we could easily add mz-compute to that set.

Real-world timings:

`touch src/storage-client/src/lib.rs; cargo check`
- Before: 9.25s
- After: 8.08s

`touch src/storage-client/src/lib.rs; cargo build -p mz-environmentd -p mz-clusterd --lib`
- Before: 45.29s
- After: 22.41s
- This is synthetic for two reasons. First, I added an empty file at src/clusterd/src/lib.rs to make --lib work. Second, as the next timings show, link times still dominate.

`touch src/storage-client/src/lib.rs; cargo build --bin environmentd --bin clusterd`
- Before: 1m40s
- After: 1m39s

Deps above mz-storage-client (before)

![deps-before](https://github.com/MaterializeInc/materialize/assets/52528/b07f14cb-6e4d-4296-be02-c3108fd81601)

Deps above mz-storage-client (after)

![deps-after](https://github.com/MaterializeInc/materialize/assets/52528/ad16da5e-2945-477b-86a1-a442da33fec1)

### Motivation

   * This PR refactors existing code.

### Tips for reviewer

Opening as a draft because there's some things to work out. Also, I got a touch sloppy with the commit hygiene and import blocks, I'll go back and tighten that up if we decide that this is desirable and if/when the final names are decided.

I feel pretty good about the foo-types model:
- There is already a mz-persist-types crate with exactly this sort of thing.
- Both mz-storage-client and mz-compute-client had types modules, which is mostly what got moved in.
- At a previous job with much worse compile times (scala), we ended up very successfully using a `foo-types`, `foo-service` (interface), `foo-concrete` model. `foo-service` corresponds to our `mz-foo-client` and `foo-concrete` corresponds to our `mz-foo`.

Names that are more or less a strawman at this point:
- I split out an mz-storage-operators to hold persist_source because it was the only reason mz-compute was depending on mz-storage-client (and we don't want to move it into mz-storage because mz-compute does not depend on mz-storage).
- As written this PR currently moves the storage controller trait out from mz-storage-client into mz-storage-types, but I think that's wrong. Instead, we should move the storage controller trait _impl_ into a new crate, something like mz-storage-controller.

Intuition for the various names and what should live in them:
- mz-storage-types, mz-compute-types: POD (plain old data), protos, no business logic (in reality, minimal business logic)
- mz-storage-client, mz-compute-client: The _adaptor_ APIs to storage and compute (aka the controllers)
- mz-storage-operators: The _compute_ API to storage
- mz-storage, mz-compute: Storage and compute dataflows

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly considered.
- [ ] This PR has an associated up-to-date [design doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md), is a design doc ([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)), or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T` mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md) (possibly in a backwards-incompatible way), then it is tagged with a `T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests, there is a companion cloud PR to account for those changes that is tagged with the release-blocker label ([example](https://github.com/MaterializeInc/cloud/pull/5021)).
  <!-- Ask in #team-cloud on Slack if you need help preparing the cloud PR. -->
- [ ] This PR includes the following [user-facing behavior changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - <!-- Add release notes here or explicitly state that there are no user-facing behavior changes. -->