@petrosagg 2023-09-08
I see. It sounds like `mz_compute` should depend on `mz_storage`, just like `mz_compute_client` depends on `mz_storage_client` but that may impact linking times. Let's leave the `mz_storage_operators` as you have it!