@danhhz 2023-09-10
TFTRs, everyone!!

I played locally with splitting the storage controller _impl_ out into a new crate instead of splitting the _trait_ out into mz-storage-types, and I do think it's better, but it's non-obvious enough that it's worth going through a round of review. I'm going to pull that out into its own PR so I can go ahead and get this one merged